# api-contract/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Baraka platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /registration-requests:
    get:
      operationId: listRegistrationRequests
      tags: [registration-requests]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RegistrationRequest"
    post:
      operationId: createRegistrationRequest
      tags: [registration-requests]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationRequest"
  /registration-requests/{id}:
    patch:
      operationId: updateRegistrationRequest
      tags: [registration-requests]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestReview"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationRequest"
...
components:
  schemas:
...
    RegistrationRequest:
      type: object
      required: [id, orgName, governorate, category, about, contactName, contactPhone, status, createdAt]
      properties:
        id: { type: integer }
        orgName: { type: string }
        governorate: { type: string }
        category: { type: string }
        regNumber: { type: ["string", "null"] }
        about: { type: string }
        contactName: { type: string }
        contactPhone: { type: string }
        email: { type: ["string", "null"] }
        status: { type: string }
        adminNote: { type: ["string", "null"] }
        reviewedBy: { type: ["integer", "null"] }
        reviewedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    RegistrationRequestInput:
      type: object
      required: [orgName, governorate, category, about, contactName, contactPhone]
      properties:
        orgName: { type: string, minLength: 2 }
        governorate: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        regNumber: { type: string }
        about: { type: string, minLength: 5 }
        contactName: { type: string, minLength: 2 }
        contactPhone: { type: string, minLength: 7 }
        email: { type: string, format: email }
    RequestReview:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [pending, approved, rejected] }
        adminNote: { type: string }
...