# api-contract/zod/src/generated/types/registrationRequestInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Baraka platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RegistrationRequestInput {
  /** @minLength 2 */
  orgName: string;
  /** @minLength 1 */
  governorate: string;
  /** @minLength 1 */
  category: string;
  regNumber?: string;
  /** @minLength 5 */
  about: string;
  /** @minLength 2 */
  contactName: string;
  /** @minLength 7 */
  contactPhone: string;
  email?: string;
}